pyproject.toml
[tool.poetry]
name = "breweries"
version = "0.1.0"
description = ""
authors = ["Ricardo Ikegame <r.ikegame@hotmail.com>"]
readme = "README.md"
packages = [
    { include = "breweries", from = "src" }
]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
requests = "^2.32.5"
pandas = "^2.3.3"
pyarrow = "^22.0.0"
fastparquet = "^2025.12.0"
deltalake = "^1.3.0"
pytest = "^9.0.2"
pytest-mock = "^3.15.1"
pytest-cov = "^7.0.0"
pandera = "^0.28.0"
apache-airflow = "2.9.3"
apache-airflow-providers-standard = ">=1.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-mock = "^3.15.1"
pytest-cov = "^7.0.0"
requests = "^2.32.5"
pandas = "^2.3.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
addopts = "--cov=src --cov-report=term-missing"
testpaths = ["tests"]
python_files = ["test_*.py"]
